refactor: start modularising into core module
