+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/arch/limits/NoopLimitStrategy.kt
@@ -1,11 +1,8 @@
 package io.embrace.android.embracesdk.internal.arch.limits
 
-import io.embrace.android.embracesdk.annotation.InternalApi
-
 /**
  * This class captures whatever the hell it wants, whenever it wants
  */
-@InternalApi
 public object NoopLimitStrategy : LimitStrategy {
 
     override fun shouldCapture(): Boolean = true
