+++ embrace-android-core/src/main/kotlin/io/embrace/android/embracesdk/internal/logging/InternalErrorHandler.kt
@@ -1,0 +1,5 @@
+package io.embrace.android.embracesdk.internal.logging
+
+public interface InternalErrorHandler {
+    public fun handleInternalError(throwable: Throwable)
+}
