--- embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/capture/internal/errors/InternalErrorHandler.kt
@@ -1,8 +1,0 @@
-package io.embrace.android.embracesdk.capture.internal.errors
-
-import io.embrace.android.embracesdk.annotation.InternalApi
-
-@InternalApi
-public interface InternalErrorHandler {
-    public fun handleInternalError(throwable: Throwable)
-}
