+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/arch/DataCaptureOrchestrator.kt
@@ -1,9 +1,9 @@
 package io.embrace.android.embracesdk.internal.arch
 
-import io.embrace.android.embracesdk.capture.internal.errors.InternalErrorType
 import io.embrace.android.embracesdk.internal.arch.datasource.DataSourceState
 import io.embrace.android.embracesdk.internal.config.ConfigService
 import io.embrace.android.embracesdk.internal.logging.EmbLogger
+import io.embrace.android.embracesdk.internal.logging.InternalErrorType
 import io.embrace.android.embracesdk.worker.BackgroundWorker
 import io.embrace.android.embracesdk.worker.TaskPriority
 import java.util.concurrent.CopyOnWriteArrayList
