+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/internal/arch/datasource/DataSourceImpl.kt
@@ -1,8 +1,8 @@
 package io.embrace.android.embracesdk.internal.arch.datasource
 
-import io.embrace.android.embracesdk.capture.internal.errors.InternalErrorType
 import io.embrace.android.embracesdk.internal.arch.limits.LimitStrategy
 import io.embrace.android.embracesdk.internal.logging.EmbLogger
+import io.embrace.android.embracesdk.internal.logging.InternalErrorType
 
 /**
  * Base class for data sources.
